refactor: extract chat window logic into dedicated component and update agent documentation for frontend best practices
